Split up lambda_utils
