+++ localstack/services/lambda_/api_utils.py
@@ -1,4 +1,6 @@
-""" Utilities for Lambda ARN handling, validations, and API output formatting."""
+""" Utilities related to Lambda API operations such as ARN handling, validations, and output formatting.
+Everything related to behavior or implicit functionality goes into `lambda_utils.py`.
+"""
 import datetime
 import random
 import re
