+++ localstack/services/lambda_/invocation/docker_runtime_executor.py
@@ -18,8 +18,8 @@ from localstack.services.lambda_.invocation.runtime_executor import (
     LambdaRuntimeException,
     RuntimeExecutor,
 )
-from localstack.services.lambda_.legacy.lambda_utils import (
-    HINT_LOG,
+from localstack.services.lambda_.lambda_utils import HINT_LOG
+from localstack.services.lambda_.networking import (
     get_all_container_networks_for_lambda,
     get_main_endpoint_from_container,
 )
