+++ localstack/services/lambda_/legacy/lambda_utils.py
@@ -1,36 +1,31 @@
 """Lambda utils for old Lambda provider"""
 import base64
-import json
 import logging
-import os
 import re
 import tempfile
 import time
 from functools import lru_cache
 from io import BytesIO
-from typing import Any, Dict, List, Optional, Union
+from typing import Any, Dict, Optional, Union
 
 from flask import Response
 
 from localstack import config
 from localstack.aws.accounts import get_aws_account_id
-from localstack.aws.api.lambda_ import FilterCriteria, Runtime
+from localstack.aws.api.lambda_ import Runtime
 from localstack.aws.connect import connect_to
-from localstack.services.lambda_.legacy.lambda_models import LambdaStoreV1, lambda_stores_v1
+from localstack.services.lambda_.legacy.lambda_models import (
+    LambdaStoreV1,
+    lambda_stores_v1,
+)
 from localstack.utils.aws import aws_stack
 from localstack.utils.aws.arns import extract_account_id_from_arn, extract_region_from_arn
 from localstack.utils.aws.aws_models import LambdaFunction
 from localstack.utils.aws.aws_responses import flask_error_response_json
-from localstack.utils.container_networking import (
-    get_endpoint_for_network,
-    get_main_container_network,
-)
 from localstack.utils.docker_utils import DOCKER_CLIENT
-from localstack.utils.strings import first_char_to_lower, short_uid
+from localstack.utils.strings import short_uid
 
 LOG = logging.getLogger(__name__)
-# Custom logger for proactive deprecation hints related to the migration from the old to the new lambda provider
-HINT_LOG = logging.getLogger("localstack.services.lambda_.hints")
 
 # root path of Lambda API endpoints
 API_PATH_ROOT = "/2015-03-31"
@@ -56,20 +51,11 @@ LAMBDA_RUNTIME_PROVIDED = Runtime.provided
 LAMBDA_RUNTIME_PROVIDED_AL2 = Runtime.provided_al2
 
 
-# default handler and runtime
-LAMBDA_DEFAULT_HANDLER = "handler.handler"
-LAMBDA_DEFAULT_RUNTIME = LAMBDA_RUNTIME_PYTHON39  # FIXME (?)
-LAMBDA_DEFAULT_STARTING_POSITION = "LATEST"
-
 # List of Dotnet Lambda runtime names
 DOTNET_LAMBDA_RUNTIMES = [
     LAMBDA_RUNTIME_DOTNETCORE31,
     LAMBDA_RUNTIME_DOTNET6,
 ]
-
-# IP address of main Docker container (lazily initialized)
-DOCKER_MAIN_CONTAINER_IP = None
-LAMBDA_CONTAINER_NETWORK = None
 
 FUNCTION_NAME_REGEX = re.compile(
     r"(arn:(aws[a-zA-Z-]*)?:lambda:)?((?P<region>[a-z]{2}(-gov)?-[a-z]+-\d{1}):)?(?P<account>\d{12}:)?(function:)?(?P<name>[a-zA-Z0-9-_\.]+)(:(?P<qualifier>\$LATEST|[a-zA-Z0-9-_]+))?"
@@ -127,37 +113,6 @@ def is_provided_runtime(runtime_details: Union[LambdaFunction, str]) -> bool:
     return runtime.startswith("provided")
 
 
-def format_name_to_path(handler_name: str, delimiter: str, extension: str):
-    file_path = handler_name.rpartition(delimiter)[0]
-    if delimiter == ":":
-        file_path = file_path.split(delimiter)[0]
-
-    if os.path.sep not in file_path:
-        file_path = file_path.replace(".", os.path.sep)
-
-    if file_path.startswith(f".{os.path.sep}"):
-        file_path = file_path[2:]
-
-    return f"{file_path}{extension}"
-
-
-def get_handler_file_from_name(handler_name: str, runtime: str = None):
-    runtime = runtime or LAMBDA_DEFAULT_RUNTIME
-
-    if runtime.startswith(LAMBDA_RUNTIME_PROVIDED):
-        return "bootstrap"
-    if runtime.startswith("nodejs"):
-        return format_name_to_path(handler_name, ".", ".js")
-    if runtime.startswith(LAMBDA_RUNTIME_GOLANG):
-        return handler_name
-    if runtime.startswith(tuple(DOTNET_LAMBDA_RUNTIMES)):
-        return format_name_to_path(handler_name, ":", ".dll")
-    if runtime.startswith("ruby"):
-        return format_name_to_path(handler_name, ".", ".rb")
-
-    return format_name_to_path(handler_name, ".", ".py")
-
-
 def is_java_lambda(lambda_details):
     runtime = getattr(lambda_details, "runtime", lambda_details)
     return runtime in [LAMBDA_RUNTIME_JAVA8, LAMBDA_RUNTIME_JAVA8_AL2, LAMBDA_RUNTIME_JAVA11]
@@ -194,26 +149,6 @@ def store_lambda_logs(
     return store_cloudwatch_logs(
         logs_client, log_group_name, log_stream_name, log_output, invocation_time
     )
-
-
-def get_main_endpoint_from_container() -> str:
-    if config.HOSTNAME_FROM_LAMBDA:
-        return config.HOSTNAME_FROM_LAMBDA
-    return get_endpoint_for_network(network=get_main_container_network_for_lambda())
-
-
-def get_main_container_network_for_lambda() -> str:
-    global LAMBDA_CONTAINER_NETWORK
-    if config.LAMBDA_DOCKER_NETWORK:
-        return config.LAMBDA_DOCKER_NETWORK.split(",")[0]
-    return get_main_container_network()
-
-
-def get_all_container_networks_for_lambda() -> list[str]:
-    global LAMBDA_CONTAINER_NETWORK
-    if config.LAMBDA_DOCKER_NETWORK:
-        return config.LAMBDA_DOCKER_NETWORK.split(",")
-    return [get_main_container_network()]
 
 
 def rm_docker_container(container_name_or_id, check_existence=False, safe=False):
@@ -315,125 +250,6 @@ def generate_lambda_arn(
         return f"arn:aws:lambda:{region}:{account_id}:function:{fn_name}"
 
 
-def parse_and_apply_numeric_filter(
-    record_value: Dict, numeric_filter: List[Union[str, int]]
-) -> bool:
-    if len(numeric_filter) % 2 > 0:
-        LOG.warning("Invalid numeric lambda filter given")
-        return True
-
-    if not isinstance(record_value, (int, float)):
-        LOG.warning(f"Record {record_value} seem not to be a valid number")
-        return False
-
-    for idx in range(0, len(numeric_filter), 2):
-        try:
-            if numeric_filter[idx] == ">" and not (record_value > float(numeric_filter[idx + 1])):
-                return False
-            if numeric_filter[idx] == ">=" and not (record_value >= float(numeric_filter[idx + 1])):
-                return False
-            if numeric_filter[idx] == "=" and not (record_value == float(numeric_filter[idx + 1])):
-                return False
-            if numeric_filter[idx] == "<" and not (record_value < float(numeric_filter[idx + 1])):
-                return False
-            if numeric_filter[idx] == "<=" and not (record_value <= float(numeric_filter[idx + 1])):
-                return False
-        except ValueError:
-            LOG.warning(
-                f"Could not convert filter value {numeric_filter[idx + 1]} to a valid number value for filtering"
-            )
-    return True
-
-
-def verify_dict_filter(record_value: any, dict_filter: Dict[str, any]) -> bool:
-    # https://docs.aws.amazon.com/lambda/latest/dg/invocation-eventfiltering.html#filtering-syntax
-    fits_filter = False
-    for key, filter_value in dict_filter.items():
-        if key.lower() == "anything-but":
-            fits_filter = record_value not in filter_value
-        elif key.lower() == "numeric":
-            fits_filter = parse_and_apply_numeric_filter(record_value, filter_value)
-        elif key.lower() == "exists":
-            fits_filter = bool(filter_value)  # exists means that the key exists in the event record
-        elif key.lower() == "prefix":
-            if not isinstance(record_value, str):
-                LOG.warning(f"Record Value {record_value} does not seem to be a valid string.")
-            fits_filter = isinstance(record_value, str) and record_value.startswith(
-                str(filter_value)
-            )
-
-        if fits_filter:
-            return True
-    return fits_filter
-
-
-def filter_stream_record(filter_rule: Dict[str, any], record: Dict[str, any]) -> bool:
-    if not filter_rule:
-        return True
-    # https://docs.aws.amazon.com/lambda/latest/dg/invocation-eventfiltering.html#filtering-syntax
-    filter_results = []
-    for key, value in filter_rule.items():
-        # check if rule exists in event
-        record_value = (
-            record.get(key.lower(), record.get(key)) if isinstance(record, Dict) else None
-        )
-        append_record = False
-        if record_value is not None:
-            # check if filter rule value is a list (leaf of rule tree) or a dict (rescursively call function)
-            if isinstance(value, list):
-                if len(value) > 0:
-                    if isinstance(value[0], (str, int)):
-                        append_record = record_value in value
-                    if isinstance(value[0], dict):
-                        append_record = verify_dict_filter(record_value, value[0])
-                else:
-                    LOG.warning(f"Empty lambda filter: {key}")
-            elif isinstance(value, dict):
-                append_record = filter_stream_record(value, record_value)
-        else:
-            # special case 'exists'
-            if isinstance(value, list) and len(value) > 0:
-                append_record = not value[0].get("exists", True)
-
-        filter_results.append(append_record)
-    return all(filter_results)
-
-
-def filter_stream_records(records, filters: List[FilterCriteria]):
-    filtered_records = []
-    for record in records:
-        for filter in filters:
-            for rule in filter["Filters"]:
-                if filter_stream_record(json.loads(rule["Pattern"]), record):
-                    filtered_records.append(record)
-                    break
-    return filtered_records
-
-
-def contains_list(filter: Dict) -> bool:
-    if isinstance(filter, dict):
-        for key, value in filter.items():
-            if isinstance(value, list) and len(value) > 0:
-                return True
-            return contains_list(value)
-    return False
-
-
-def validate_filters(filter: FilterCriteria) -> bool:
-    # filter needs to be json serializeable
-    for rule in filter["Filters"]:
-        try:
-            if not (filter_pattern := json.loads(rule["Pattern"])):
-                return False
-            return contains_list(filter_pattern)
-        except json.JSONDecodeError:
-            return False
-    # needs to contain on what to filter (some list with citerias)
-    # https://docs.aws.amazon.com/lambda/latest/dg/invocation-eventfiltering.html#filtering-syntax
-
-    return True
-
-
 def function_name_from_arn(arn: str):
     """Extract a function name from a arn/function name"""
     return FUNCTION_NAME_REGEX.match(arn).group("name")
@@ -457,14 +273,3 @@ def get_lambda_store_v1_for_arn(resource_arn: str) -> LambdaStoreV1:
         account_id=extract_account_id_from_arn(resource_arn or ""),
         region=extract_region_from_arn(resource_arn or ""),
     )
-
-
-def message_attributes_to_lower(message_attrs):
-    """Convert message attribute details (first characters) to lower case (e.g., stringValue, dataType)."""
-    message_attrs = message_attrs or {}
-    for _, attr in message_attrs.items():
-        if not isinstance(attr, dict):
-            continue
-        for key, value in dict(attr).items():
-            attr[first_char_to_lower(key)] = attr.pop(key)
-    return message_attrs
