+++ localstack/services/lambda_/provider.py
@@ -141,6 +141,7 @@ from localstack.services.lambda_.api_utils import ARCHITECTURES, STATEMENT_ID_RE
 from localstack.services.lambda_.event_source_listeners.event_source_listener import (
     EventSourceListener,
 )
+from localstack.services.lambda_.event_source_listeners.utils import validate_filters
 from localstack.services.lambda_.invocation import AccessDeniedException
 from localstack.services.lambda_.invocation.execution_environment import (
     EnvironmentStartupTimeoutException,
@@ -183,7 +184,6 @@ from localstack.services.lambda_.invocation.lambda_service import (
 from localstack.services.lambda_.invocation.models import LambdaStore
 from localstack.services.lambda_.invocation.runtime_executor import get_runtime_executor
 from localstack.services.lambda_.layerfetcher.layer_fetcher import LayerFetcher
-from localstack.services.lambda_.legacy.lambda_utils import validate_filters
 from localstack.services.lambda_.urlrouter import FunctionUrlRouter
 from localstack.services.plugins import ServiceLifecycleHook
 from localstack.state import StateVisitor
