+++ tests/aws/services/apigateway/test_apigateway_lambda.py
@@ -7,7 +7,6 @@ from botocore.exceptions import ClientError
 
 from localstack.aws.api.lambda_ import Runtime
 from localstack.constants import TEST_AWS_REGION_NAME
-from localstack.services.lambda_.legacy.lambda_utils import LAMBDA_RUNTIME_PYTHON39
 from localstack.testing.pytest import markers
 from localstack.utils.aws import arns
 from localstack.utils.files import load_file
@@ -113,7 +112,7 @@ def test_lambda_aws_proxy_integration(
         func_name=function_name,
         handler_file=TEST_LAMBDA_AWS_PROXY,
         handler="lambda_aws_proxy.handler",
-        runtime=LAMBDA_RUNTIME_PYTHON39,
+        runtime=Runtime.python3_9,
     )
     # create invocation role
     _, role_arn = create_role_with_policy(
@@ -265,7 +264,7 @@ def test_lambda_aws_integration(
         func_name=function_name,
         handler_file=TEST_LAMBDA_PYTHON_ECHO,
         handler="lambda_echo.handler",
-        runtime=LAMBDA_RUNTIME_PYTHON39,
+        runtime=Runtime.python3_9,
     )
     # create invocation role
     _, role_arn = create_role_with_policy(
@@ -347,7 +346,7 @@ def test_lambda_aws_integration_with_request_template(
         func_name=function_name,
         handler_file=TEST_LAMBDA_PYTHON_ECHO,
         handler="lambda_echo.handler",
-        runtime=LAMBDA_RUNTIME_PYTHON39,
+        runtime=Runtime.python3_9,
     )
     # create invocation role
     _, role_arn = create_role_with_policy(
@@ -447,7 +446,7 @@ def test_lambda_aws_integration_response_with_mapping_templates(
         func_name=function_name,
         handler_file=TEST_LAMBDA_MAPPING_RESPONSES,
         handler="lambda_mapping_responses.handler",
-        runtime=LAMBDA_RUNTIME_PYTHON39,
+        runtime=Runtime.python3_9,
     )
     # create invocation role
     _, role_arn = create_role_with_policy(
